Move QueuedStatementResource helpers to QueryResourceUtil
